backend/requirements.txt
celery==5.3.6
redis==5.0.4
pytest==8.2.2
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.25
pymysql==1.1.0
requests==2.32.3
python-dotenv==1.0.1
markdown2==2.5.0
beautifulsoup4==4.12.3
lxml[html_clean]==5.3.0
readability-lxml==0.8.1
croniter==1.4.1
playwright==1.49.0
crawl4ai>=0.7.0
alibabacloud_iqs20241111==1.6.2
alibabacloud_tea_openapi==0.4.2
alibabacloud_tea_util==0.3.13
oss2==2.18.6